Rename runtime files to match functions
